+++ src/main/java/com/buuz135/transfer_labels/item/TransferLabelItem.java
@@ -38,7 +38,7 @@ public abstract class TransferLabelItem extends Item {
         if (context.getLevel() instanceof ServerLevel serverLevel) {
             LabelStorage.addLabel(serverLevel, context.getClickedPos(), context.getClickedFace(), context.getItemInHand().copyWithCount(1));
             context.getItemInHand().shrink(1);
-            TransferLabels.NETWORK.sendTo(new LabelSyncPacket(serverLevel.getLevel().dimension().location(), LabelStorage.getStorageFor(serverLevel).saveNearby(context.getPlayer().getOnPos(), 16, serverLevel.registryAccess()), context.getPlayer().getOnPos(), 16), (ServerPlayer) context.getPlayer());
+            TransferLabels.NETWORK.sendTo(new LabelSyncPacket(LabelStorage.getStorageFor(serverLevel).saveNearby(context.getPlayer().getOnPos(), 100, serverLevel.registryAccess()), context.getPlayer().getOnPos(), 16), (ServerPlayer) context.getPlayer());
             return InteractionResult.SUCCESS;
         }